+++ setup.cfg
@@ -26,7 +26,7 @@ pretty = True
 
 # the __init__ files have dynamic check - ignoring the attribute error. others are generated files
 # top level package name only ignores the __init__.py file.
-[mypy-xonsh.parser_table,xonsh.completion_parser_table,xonsh.parsers.parser_table.*,xonsh.parsers.completion_parser_table.*,xonsh,xonsh.prompt,xonsh.history,xonsh.completers,xonsh.procs]
+[mypy-xonsh.parser_table,xonsh.completion_parser_table,xonsh.parser.parser_table.*,xonsh.parser.completion_parser_table.*,xonsh,xonsh.prompt,xonsh.history,xonsh.completers,xonsh.procs]
 ignore_errors = True
 
 # 3rd party libraries that we dont have control over
