+++ tests/completers/test_base_completer.py
@@ -1,7 +1,7 @@
 import pytest
 
 from xonsh.completers.base import complete_base
-from xonsh.parsers.completion_context import CommandContext, CompletionContext
+from xonsh.parser.completion_context import CommandContext, CompletionContext
 from xonsh.pytest.tools import ON_WINDOWS
 
 CUR_DIR = (
