+++ tests/parser/test_parser.py
@@ -8,7 +8,7 @@ import pytest
 
 from xonsh.ast import AST, Call, Pass, With, is_const_str
 from xonsh.parser import Parser
-from xonsh.parsers.fstring_adaptor import FStringAdaptor
+from xonsh.parser.fstring_adaptor import FStringAdaptor
 from xonsh.pytest.tools import (
     VER_MAJOR_MINOR,
     nodes_equal,
