+++ xompletions/pip.py
@@ -1,7 +1,7 @@
 """Completers for pip."""
 
 from xonsh.completers.tools import comp_based_completer
-from xonsh.parsers.completion_context import CommandContext
+from xonsh.parser.completion_context import CommandContext
 
 
 def xonsh_complete(ctx: CommandContext):
