+++ xonsh/parser/base.py
@@ -14,8 +14,8 @@ from xonsh import ast
 from xonsh.ast import has_elts, load_attribute_chain, xonsh_call
 from xonsh.lazyasd import LazyObject
 from xonsh.lexer import Lexer, LexToken
-from xonsh.parsers.context_check import check_contexts
-from xonsh.parsers.fstring_adaptor import FStringAdaptor
+from xonsh.parser.context_check import check_contexts
+from xonsh.parser.fstring_adaptor import FStringAdaptor
 from xonsh.platform import PYTHON_VERSION_INFO
 from xonsh.ply.ply import yacc
 from xonsh.tokenize import SearchPath, StringPrefix
