+++ xonsh/parser/v36.py
@@ -1,7 +1,7 @@
 """Implements the xonsh parser for Python v3.6."""
 
 import xonsh.ast as ast
-from xonsh.parsers.base import BaseParser, lopen_loc, store_ctx
+from xonsh.parser.base import BaseParser, lopen_loc, store_ctx
 
 
 class Parser(BaseParser):
