+++ xonsh/parser/v38.py
@@ -1,8 +1,8 @@
 """Implements the xonsh parser for Python v3.8."""
 
 import xonsh.ast as ast
-from xonsh.parsers.base import store_ctx
-from xonsh.parsers.v36 import Parser as ThreeSixParser
+from xonsh.parser.base import store_ctx
+from xonsh.parser.v36 import Parser as ThreeSixParser
 
 
 class Parser(ThreeSixParser):
